+++ CHANGELOG.md
@@ -70,6 +70,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (evm) [#184](https://github.com/EscanBE/evermint/pull/184) Warmup coinbase + custom-precompiled-contracts and fix some tests
 - (cpc) [#186](https://github.com/EscanBE/evermint/pull/186) Correct genesis import/export for `x/cpc`
 - (rename-chain) [#189](https://github.com/EscanBE/evermint/pull/189) Fix broken feature rename-chain
+- (cpc) [#191](https://github.com/EscanBE/evermint/pull/191) Add pseudocode when query `eth_getCode` and gRPC query code for custom precompiled contracts
 
 ### API Breaking